made some pretty broad strokes to rearchitecture the export to just put any csv related content into csv.rs file
